combined build and test recipes into single cmake_recipe module
