+++ cpm/domain/build_service.py
@@ -2,7 +2,7 @@ class BuildService(object):
     def __init__(self, project_loader):
         self.project_loader = project_loader
 
-    def build(self, compilation_recipe):
+    def build(self, cmake_recipe):
         project = self.project_loader.load()
-        compilation_recipe.generate(project)
-        compilation_recipe.compile(project)
+        cmake_recipe.generate(project)
+        cmake_recipe.build(project)
