--- cpm/domain/compilation_recipes/__init__.py
@@ -1,13 +1,0 @@
-RECIPES_DIRECTORY = 'recipes'
-
-
-class CompilationRecipe(object):
-    def generate(self, project):
-        raise NotImplementedError()
-
-    def compile(self, project):
-        raise NotImplementedError()
-
-
-class CompilationError(RuntimeError):
-    pass
